+++ src/main/java/com/site/eterroir/service/implementation/CooperativeServiceImpl.java
@@ -1,20 +1,23 @@
 package com.site.eterroir.service.implementation;
 
 import com.site.eterroir.model.Cooperative;
+import com.site.eterroir.model.Cooperative;
 import com.site.eterroir.repository.CooperativeRepo;
+import com.site.eterroir.security.SecurityService;
 import com.site.eterroir.service.CooperativeService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
 import java.util.List;
+import java.util.Map;
 
 @RequiredArgsConstructor
 @Service
 @Transactional
-@Slf4j
 public class CooperativeServiceImpl implements CooperativeService {
 
     private final CooperativeRepo cooperativeRepo;
@@ -22,9 +25,12 @@ public class CooperativeServiceImpl implements CooperativeService {
 
     @Override
     public Cooperative create(Cooperative cooperative) {
-        log.info("saving new cooperative: " + cooperative.getEmail());
-        cooperative.setMotDePasse(passwordEncoder.encode(cooperative.getMotDePasse()));
-        return cooperativeRepo.save(cooperative);
+        if(SecurityService.isAdmin()){
+            cooperative.setMotDePasse(passwordEncoder.encode(cooperative.getMotDePasse()));
+            return cooperativeRepo.save(cooperative);
+        } else {
+            throw new AccessDeniedException("Vous n'avez pas l'autorisation");
+        }
     }
 
     @Override
@@ -39,18 +45,46 @@ public class CooperativeServiceImpl implements CooperativeService {
 
     @Override
     public Boolean delete(Long id) {
-        cooperativeRepo.deleteById(id);
-        return true;
+        if(SecurityService.isAdmin()){
+            cooperativeRepo.deleteById(id);
+            return true;
+        } else {
+            throw new AccessDeniedException("Vous n'avez pas l'autorisation");
+        }
     }
 
     @Override
-    public Cooperative update(Long id, Cooperative cooperative) throws Exception {
-        if(cooperativeRepo.findById(id).isPresent()) {
-            cooperative.setId(id);
-            return cooperativeRepo.save(cooperative);
-        } else{
-            throw new Exception("Id n'existe pas");
+    public Cooperative update(Long id, Cooperative cooperative){
+        Cooperative dbCooperative = cooperativeRepo.findById(id).get();
+        if(SecurityService.isAdmin() || isOwner(dbCooperative)) {
+            if(cooperative.getEmail() != null)
+                dbCooperative.setEmail(cooperative.getEmail());
+            if(cooperative.getMotDePasse() != null)
+                dbCooperative.setMotDePasse(passwordEncoder.encode(cooperative.getMotDePasse()));
+            if(cooperative.getNumTele() != null)
+                dbCooperative.setNumTele(cooperative.getNumTele());
+            if(cooperative.getNom() != null)
+                dbCooperative.setNom(cooperative.getNom());
+            if(cooperative.getSecteur() != null)
+                dbCooperative.setSecteur(cooperative.getSecteur());
+            if(cooperative.getRegion() != null)
+                dbCooperative.setRegion(cooperative.getRegion());
+            return cooperativeRepo.save(dbCooperative);
+        } else {
+            throw new AccessDeniedException("Vous n'avez pas l'autorisation");
         }
+    }
+
+
+    // User check methods:
+
+    private boolean isOwner(Cooperative cooperative) {
+        Map<String, String> user = SecurityService.getLoggedUser();
+    if( user.get("authority").contains("COOPERATIVE_ROLE")
+                &&  user.get("email").equals(cooperative.getEmail()) ) {
+            return true;
+        }
+        return false;
     }
 
 }
