+++ src/main/java/com/site/eterroir/service/implementation/MatierePremiereServiceImpl.java
@@ -3,8 +3,10 @@ package com.site.eterroir.service.implementation;
 import com.site.eterroir.dto.MatierePremiereDto;
 import com.site.eterroir.model.MatierePremiere;
 import com.site.eterroir.repository.MatierePremiereRepo;
+import com.site.eterroir.security.SecurityService;
 import com.site.eterroir.service.MatierePremiereService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -20,8 +22,12 @@ public class MatierePremiereServiceImpl implements MatierePremiereService {
 
     @Override
     public MatierePremiereDto create(MatierePremiereDto matiereDto) {
-        MatierePremiere matiere = convertToEntity(matiereDto);
-        return convertToDto(matiereRepo.save(matiere));
+        if(SecurityService.isAdmin()) {
+            MatierePremiere matiere = dtoToEntity(matiereDto);
+            return entityToDto(matiereRepo.save(matiere));
+        } else {
+            throw new AccessDeniedException("Vous n'avez pas l'autorisation");
+        }
     }
 
     @Override
@@ -29,37 +35,43 @@ public class MatierePremiereServiceImpl implements MatierePremiereService {
         List<MatierePremiere> matieres = matiereRepo.findAll();
         List<MatierePremiereDto> dtoList = new ArrayList<>();
         for(MatierePremiere matiere : matieres){
-            dtoList.add(convertToDto(matiere));
+            dtoList.add(entityToDto(matiere));
         }
         return dtoList;
     }
 
     @Override
     public MatierePremiereDto get(Long id) {
-        return convertToDto(matiereRepo.findById(id).get());
+        return entityToDto(matiereRepo.findById(id).get());
     }
 
     @Override
     public Boolean delete(Long id) {
-        matiereRepo.deleteById(id);
-        return true;
+        if(SecurityService.isAdmin()) {
+            matiereRepo.deleteById(id);
+            return true;
+        } else {
+            throw new AccessDeniedException("Vous n'avez pas l'autorisation");
+        }
     }
 
     @Override
-    public MatierePremiereDto update(Long id, MatierePremiereDto matiereDto) throws Exception {
-        if(matiereRepo.findById(id).isPresent()) {
-            MatierePremiere matiere = convertToEntity(matiereDto);
-            matiere.setId(id);
-            return convertToDto(matiereRepo.save(matiere));
-        } else{
-            throw new Exception("Id n'existe pas");
+    public MatierePremiereDto update(Long id, MatierePremiereDto matiereDto) {
+        if(SecurityService.isAdmin()) {
+            MatierePremiere dbMatierePremiere = matiereRepo.findById(id).get();
+            MatierePremiere matiere = dtoToEntity(matiereDto);
+            if(matiere.getNom() != null)
+                dbMatierePremiere.setNom(matiere.getNom());
+            return entityToDto(matiereRepo.save(dbMatierePremiere));
+        } else {
+            throw new AccessDeniedException("Vous n'avez pas l'autorisation");
         }
     }
 
 
     // DTO conversion:
 
-    private MatierePremiereDto convertToDto(MatierePremiere matiere){
+    private MatierePremiereDto entityToDto(MatierePremiere matiere){
         MatierePremiereDto matiereDto = new MatierePremiereDto();
         matiereDto.setId(matiere.getId());
         matiereDto.setNom(matiere.getNom());
@@ -67,13 +79,12 @@ public class MatierePremiereServiceImpl implements MatierePremiereService {
         return matiereDto;
     };
 
-    private MatierePremiere convertToEntity(MatierePremiereDto matiereDto){
+    private MatierePremiere dtoToEntity(MatierePremiereDto matiereDto){
         MatierePremiere matiere = new MatierePremiere();
         matiere.setId(matiereDto.getId());
         matiere.setNom(matiereDto.getNom());
 
         return matiere;
     };
-
 
 }
